Remove public key requirement to decrypt
OpenSSL::PKCS7.decrypt validates the recipient by comparing the
serial number of the recipient certificate with the one bundled
with the data. It also makes sure the public keys match. Since,
the serial number is bundled with the data and the public key
is bundled with the private key, we can generate on the fly
a certificate object that satisfies PKCS7.decrypt and return
the plain text.